Create Winner class and utilize in determineWinner
Previously the determineWinner function was only returning whether
the player or the computer won but did not mention either choices.
Utilizing a Winner class which stores who won and a message can
accomplish this goal.

Coming from a java background, a point of confusion for me while
experimenting with javascript is where to place a new object.