Delete unnecessary comments
Though the pseudocode before some of the functions were useful while
I was writing the code, I do not want to include unnecessary comments
that make the code difficult to read. I think that the simple one line
explanations before each function are sufficient and the code itself
is clear enough.