Refactor game logic to work with buttons
Changed buttonClicked function to playRound and added the game logic
into the playRound function which is called when a button is clicked.

I was struggling with figuring out how to get the game to stop when
the playerScore or computerScore was 3, like how it was with the while
loop.

I tried putting the DOM manipulation logic inside the while loop, which
ended up crashing my website. I was trying to figure out a simple way to
insert the button functionality into my pre-existing code and just have
the button return a playerSelection but this proved very difficult.

I ended up having to just check for the playerScore and computerScore
each time a button was clicked.